% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sr_as_df}
\alias{sr_as_df}
\title{Converting Stadtrad Bike Positions to Dataframe (Long Format)}
\usage{
sr_as_df(x, ...)
}
\arguments{
\item{x}{Object of class \code{srbike} as returned by \code{\link[=sr_load_data]{sr_load_data()}}.}
}
\description{
Converting Stadtrad Bike Positions to Dataframe (Long Format)
}
\author{
Reto
}
