% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sr_connect}
\alias{sr_connect}
\title{Connecting to Existing Stadtrad Database}
\usage{
sr_connect(file)
}
\arguments{
\item{file}{Path to the sqlite3 database.}
}
\value{
Returns an object of class \code{SQLiteConnection}.
}
\description{
Connects to sqlite3 database and checks if the database
is as expected before returning a database connection.
}
\author{
Reto
}
